requirements.txt
fastapi[standard]==0.103.2
uvicorn[standard]==0.22.0
pymysql==1.1.1
python-dotenv==0.21.1
sqlalchemy==2.0.40
boto3==1.37.23
